Rename mouse file to input
